# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /agents:
    get:
      operationId: getAgents
      tags: [agents]
      summary: List agent manifests
      description: >-
        Returns agent manifest metadata and skill-reference readiness without
        exposing secrets or configured absolute local paths.
      parameters:
        - name: teamId
          in: query
          required: false
          schema:
            type: string
          description: Optional team id filter.
        - name: runtimeStatus
          in: query
          required: false
          schema:
            type: string
            enum:
              - runnable
              - template
          description: Optional runtime status filter.
      responses:
        "200":
          description: Agent manifests and readiness
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentListResponse"
        "400":
          description: Invalid query filter
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /agents/{agentId}/export/vscode-agent:
    get:
      operationId: exportVscodeAgent
      tags: [agents]
      summary: Export a VS Code agent file
      description: >-
        Returns a VS Code-compatible `.agent.md` file with YAML front matter
        containing the agent description and registered bound skill IDs as
        tools. Secret-looking values, local provider URLs, MCP server URLs, and
        local paths are redacted from exported text.
      parameters:
        - name: agentId
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/AgentId"
      responses:
        "200":
          description: VS Code agent Markdown
          content:
            text/markdown:
              schema:
                type: string
        "404":
          description: Agent not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /agents/{agentId}/export/mcp-tool:
    get:
      operationId: exportAgentMcpTool
      tags: [agents]
      summary: Export MCP wrapper tool metadata
      description: >-
        Returns redacted metadata for exposing an agent run entrypoint as an
        MCP-callable tool. Provider internals and local configuration values are
        omitted.
      parameters:
        - name: agentId
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/AgentId"
      responses:
        "200":
          description: MCP wrapper tool metadata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentMcpToolMetadata"
        "404":
          description: Agent not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    SkillId:
      type: string
      minLength: 1
...
    AgentId:
      type: string
      pattern: "^[a-z][a-z0-9_]{1,63}$"
    AgentSource:
      type: string
      enum:
        - builtin
        - community
        - custom
        - external
    AgentMemoryPromotionMode:
      type: string
      enum:
        - manual
        - run_summary
        - disabled
    AgentSkillBinding:
      type: object
      properties:
        skillId:
          $ref: "#/components/schemas/SkillId"
        required:
          type: boolean
      required:
        - skillId
        - required
    AgentHandoff:
      type: object
      properties:
        targetAgentId:
          $ref: "#/components/schemas/AgentId"
        description:
          type: string
      required:
        - targetAgentId
        - description
    AgentManifestTest:
      type: object
      properties:
        name:
          type: string
        prompt:
          type: string
        expectedSkillIds:
          type: array
          items:
            $ref: "#/components/schemas/SkillId"
      required:
        - name
        - prompt
        - expectedSkillIds
    AgentCriticalRule:
      type: object
      properties:
        id:
          type: string
        description:
          type: string
        severity:
          type: string
          enum:
            - blocker
            - warning
      required:
        - id
        - description
        - severity
    AgentDeliverable:
      type: object
      properties:
        name:
          type: string
        format:
          type: string
        description:
          type: string
        successCriteria:
          type: string
      required:
        - name
        - format
        - description
        - successCriteria
    AgentWorkflowPhase:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        approvalRequired:
          type: boolean
        deliverables:
          type: array
          items:
            type: string
      required:
        - name
        - description
        - approvalRequired
        - deliverables
    AgentSuccessMetric:
      type: object
      properties:
        metric:
          type: string
        target:
          type: string
        measurement:
          type: string
      required:
        - metric
        - target
        - measurement
    AgentManifest:
      type: object
      properties:
        agentId:
          $ref: "#/components/schemas/AgentId"
        name:
          type: string
        title:
          type: string
        description:
          type: string
        source:
          $ref: "#/components/schemas/AgentSource"
        instructions:
          type: string
        skills:
          type: array
          items:
            $ref: "#/components/schemas/AgentSkillBinding"
        provider:
          type: object
          properties:
            provider:
              $ref: "#/components/schemas/AgentProvider"
            modelId:
              type: string
            reasoningEffort:
              $ref: "#/components/schemas/AgentReasoningEffort"
        planner:
          type: object
          properties:
            mode:
              $ref: "#/components/schemas/AgentRuntimePlanMode"
            failureStrategy:
              $ref: "#/components/schemas/DagFailureStrategy"
          required:
            - mode
            - failureStrategy
        permissions:
          type: object
          properties:
            approvalRequired:
              type: boolean
            canUseNetwork:
              type: boolean
            canWriteDatabase:
              type: boolean
          required:
            - approvalRequired
            - canUseNetwork
            - canWriteDatabase
        memory:
          type: object
          properties:
            promotionMode:
              $ref: "#/components/schemas/AgentMemoryPromotionMode"
          required:
            - promotionMode
        handoffs:
          type: array
          items:
            $ref: "#/components/schemas/AgentHandoff"
        tests:
          type: array
          items:
            $ref: "#/components/schemas/AgentManifestTest"
        identity:
          type: object
          properties:
            persona:
              type: string
            background:
              type: string
          required:
            - persona
            - background
        criticalRules:
          type: array
          items:
            $ref: "#/components/schemas/AgentCriticalRule"
        deliverables:
          type: array
          items:
            $ref: "#/components/schemas/AgentDeliverable"
        workflow:
          type: array
          items:
            $ref: "#/components/schemas/AgentWorkflowPhase"
        communicationStyle:
          type: object
          properties:
            tone:
              type: string
            outputFormat:
              type: string
            languagePreference:
              type: string
          required:
            - tone
            - outputFormat
            - languagePreference
        successMetrics:
          type: array
          items:
            $ref: "#/components/schemas/AgentSuccessMetric"
        teamId:
          type: string
        runtimeStatus:
          type: string
          default: runnable
          enum:
            - runnable
            - template
      required:
        - agentId
        - name
        - description
        - source
        - instructions
        - skills
        - planner
        - permissions
        - memory
        - handoffs
        - tests
    AgentReadiness:
      type: object
      properties:
        agentId:
          $ref: "#/components/schemas/AgentId"
        status:
          type: string
          enum:
            - ready
            - missing_skills
        missingSkillIds:
          type: array
          items:
            $ref: "#/components/schemas/SkillId"
        enabledSkillIds:
          type: array
          items:
            $ref: "#/components/schemas/SkillId"
      required:
        - agentId
        - status
        - missingSkillIds
        - enabledSkillIds
    AgentListResponse:
      type: object
      properties:
        agents:
          type: array
          items:
            $ref: "#/components/schemas/AgentManifest"
        readiness:
          type: array
          items:
            $ref: "#/components/schemas/AgentReadiness"
      required:
        - agents
        - readiness
...
    AgentMcpToolInputSchema:
      type: object
      properties:
        type:
          type: string
          const: object
        properties:
          type: object
          properties:
            message:
              type: object
              properties:
                type:
                  type: string
                  const: string
                description:
                  type: string
              required:
                - type
                - description
            executionMode:
              type: object
              properties:
                type:
                  type: string
                  const: string
                enum:
                  type: array
                  items:
                    $ref: "#/components/schemas/AgentRunExecutionMode"
                description:
                  type: string
              required:
                - type
                - enum
                - description
          required:
            - message
            - executionMode
        required:
          type: array
          items:
            type: string
          minItems: 1
        additionalProperties:
          type: boolean
          const: false
      required:
        - type
        - properties
        - required
        - additionalProperties
    AgentMcpToolMetadata:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        inputSchema:
          $ref: "#/components/schemas/AgentMcpToolInputSchema"
      required:
        - name
        - description
        - inputSchema
...
    AgentProvider:
      type: string
      enum:
        - openai
        - anthropic
        - ollama
        - deterministic
...
    AgentReasoningEffort:
      type: string
      enum:
        - none
        - low
        - medium
        - high
        - xhigh
...
    AgentRunExecutionMode:
      type: string
      enum:
        - plan_only
        - execute_ready
    AgentRuntimePlanMode:
      type: string
      enum:
        - linear
        - dag
    DagFailureStrategy:
      type: string
      enum:
        - fail_fast
        - continue_independent
...